job-description/pyproject.toml
[project]
name = "job-description"
version = "0.1.0"
description = "AI-powered job description generator using Mistral and Anthropic Claude"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic-ai==0.1.6",
    "pydantic>=2.0.0",
    "anthropic>=0.40.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "python-docx>=1.0.0",
]

[project.scripts]
job-description = "job_description.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
